pyproject.toml
[project]
name = "vapi-rental"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
